+++ src/ncgga_engine.py
@@ -6,10 +6,9 @@ from pyscf import scf
 import pyscf.scf.cphf
 import os
 
-MAXMEM = int(os.getenv("MAXMEM", 2))
+MAXMEM = float(os.getenv("MAXMEM", 2))
 np.einsum = partial(np.einsum, optimize=["greedy", 1024 ** 3 * MAXMEM / 8])
 np.set_printoptions(8, linewidth=1000, suppress=True)
-
 
 class NCGGAEngine:
